refactor: implement filter widgets as child of multiselect widget & make code more pythonic
